pep8 and pyflakes are ruthless
Since this is an older project for me,
I decided to try to fix any stylisation errors
BOY WAS THAT A BAD IDEA
The way this program works is a whole lotta `if`
statements and it means W0401 is going to be
ignored from now on. I'll try to stay in-line
but you can only indent so much in python.

Hopefully whoever reads this is not too angry at me.